+++ weather/convert.py
@@ -5,11 +5,10 @@ from netCDF4 import Dataset, num2date
 from pypsdm.db.weather.models import WeatherValue
 from sqlmodel import Session
 
-# Correct base time: 1970-01-01 for "seconds since 1970-01-01"
 BASE_TIME = datetime(year=1970, month=1, day=1, tzinfo=pytz.utc)
 
 
-def convert(
+def convert_netCFD(
     session: Session,
     accum_data: Dataset,
     instant_data: Dataset,
@@ -56,16 +55,17 @@ def convert(
 
             fDir = float(fdir_array[time_idx, lat_idx, lon_idx])
             influx_total = float(ssrd_array[time_idx, lat_idx, lon_idx])
+            time = (datetime.strptime(time_value, "%Y-%m-%dT%H:%M:%SZ"),)
 
             # Create WeatherValue object
-            weather_value = WeatherValue(
-                time=datetime.strptime(time_value, "%Y-%m-%dT%H:%M:%SZ"),
+            weather_value = make_weather_value(
+                time=time,
                 coordinate_id=coordinate_id,
-                aswdifd_s=(influx_total - fDir) / 3600,  # Difference of ssdr - fDir
-                aswdir_s=fDir / 3600,  # J/m² in Wh/m²
-                t2m=temp,
-                u131m=u131m,
-                v131m=v131m,
+                ssrd=influx_total,
+                fdir=fDir,
+                temp=temp,
+                u_wind=u131m,
+                v_wind=v131m,
             )
 
             weather_values.append(weather_value)
@@ -87,3 +87,26 @@ def convert(
         print(
             f"Final commit: {len(weather_values)} records. Total processed: {total_records}"
         )
+
+
+def make_weather_value(time, coordinate_id, ssrd, fdir, temp, u_wind, v_wind):
+    """
+    Helper to create a WeatherValue instance from weather parameters.
+        Args:
+        time: time of the weather data
+        coordinate_id: ID of the coordinate of the weather data
+        ssrd: the total influx
+        fdir: the direct influx
+        temp: the temperature at 2m
+        u_wind: the u-component of the wind velocity
+        v_wind: the v-component of the wind velocity
+    """
+    return WeatherValue(
+        time=time,
+        coordinate_id=coordinate_id,
+        aswdifd_s=(ssrd - fdir) / 3600,  # Diffuse radiation (J/m² to Wh/m²)
+        aswdir_s=fdir / 3600,  # Direct radiation (J/m² to Wh/m²)
+        t2m=temp,
+        u131m=u_wind,
+        v131m=v_wind,
+    )